+++ pkg/inst/tmvs_quantity.m
@@ -70,7 +70,7 @@ if ischar (x)
     error (sprintf ('physical quantity ''%s'' not known', x));
   end
 
-  y = int8 (y);
+  y = uint32 (y);
 elseif isindex (x)
   switch x
   case 1
