+++ pkg/inst/tmvs_section.m
@@ -22,7 +22,7 @@ if ischar (x)
     error (sprintf ('section ''%s'' not known', x));
   end
 
-  y = int8 (y);
+  y = uint32 (y);
 elseif isindex (x)
   switch x
   case 1
